initial import of the Less Than Simple DynDNS Updater
This is the initial draft of a working(-ish) DynDNS update client, in python. Using only Python's StdLib. It's about as simple as it really needs to be.

The config file is verified working.

It's currently built to get interface information from BSD-like ifconfig output. I'll eventually throw in Linux support as well.